refactor: centralise safe mode card UI
